sonar fix

+++ infrastructure-provisioning/src/general/lib/aws/actions_lib.py
@@ -523,22 +523,21 @@ def remove_all_iam_resources(instance_type, scientist=''):
         if roles_list:
             roles_list.sort(reverse=True)
             for iam_role in roles_list:
-                if '-ssn-Role' in iam_role:
-                    if instance_type == 'ssn' or instance_type == 'all':
-                        try:
-                            client.delete_role_policy(RoleName=iam_role, PolicyName=service_base_name + '-ssn-Policy')
-                        except:
-                            print('There is no policy {}-ssn-Policy to delete'.format(service_base_name))
-                        role_profiles = client.list_instance_profiles_for_role(RoleName=iam_role).get('InstanceProfiles')
-                        if role_profiles:
-                            for i in role_profiles:
-                                role_profile_name = i.get('InstanceProfileName')
-                                if role_profile_name == service_base_name + '-ssn-Profile':
-                                    remove_roles_and_profiles(iam_role, role_profile_name)
-                        else:
-                            print("There is no instance profile for {}".format(iam_role))
-                            client.delete_role(RoleName=iam_role)
-                            print("The IAM role {} has been deleted successfully".format(iam_role))
+                if '-ssn-Role' in iam_role and instance_type == 'ssn' or instance_type == 'all':
+                    try:
+                        client.delete_role_policy(RoleName=iam_role, PolicyName=service_base_name + '-ssn-Policy')
+                    except:
+                        print('There is no policy {}-ssn-Policy to delete'.format(service_base_name))
+                    role_profiles = client.list_instance_profiles_for_role(RoleName=iam_role).get('InstanceProfiles')
+                    if role_profiles:
+                        for i in role_profiles:
+                            role_profile_name = i.get('InstanceProfileName')
+                            if role_profile_name == service_base_name + '-ssn-Profile':
+                                remove_roles_and_profiles(iam_role, role_profile_name)
+                    else:
+                        print("There is no instance profile for {}".format(iam_role))
+                        client.delete_role(RoleName=iam_role)
+                        print("The IAM role {} has been deleted successfully".format(iam_role))
                 if '-edge-Role' in iam_role:
                     if instance_type == 'edge' and scientist in iam_role:
                         remove_detach_iam_policies(iam_role, 'delete')
@@ -591,10 +590,9 @@ def remove_all_iam_resources(instance_type, scientist=''):
                 profile_list.append(item.get('InstanceProfileName'))
         if profile_list:
             for instance_profile in profile_list:
-                if '-ssn-Profile' in instance_profile:
-                    if instance_type == 'ssn' or instance_type == 'all':
-                        client.delete_instance_profile(InstanceProfileName=instance_profile)
-                        print("The instance profile {} has been deleted successfully".format(instance_profile))
+                if '-ssn-Profile' in instance_profile and instance_type == 'ssn' or instance_type == 'all':
+                    client.delete_instance_profile(InstanceProfileName=instance_profile)
+                    print("The instance profile {} has been deleted successfully".format(instance_profile))
                 if '-edge-Profile' in instance_profile:
                     if instance_type == 'edge' and scientist in instance_profile:
                         client.delete_instance_profile(InstanceProfileName=instance_profile)
